refactor: 도메인 중심 구조로 패키지 변경 및 리뷰 관련 에러 처리 구현
- 도메인을 최상단 패키지로 두도록 변경
- 점수 예외 및 중복 리뷰 방지 관련 커스텀 예외 구현
- validator에 유효성 검증 구현
