+++ src/main/java/com/reboot_course/review_service/domain/product/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package com.reboot_course.review_service.domain.product.repository;
+
+import com.reboot_course.review_service.domain.product.entity.Product;
+import org.springframework.data.repository.CrudRepository;
+
+public interface ProductRepository extends CrudRepository<Product, Long> {
+}
