+++ src/main/java/com/reboot_course/review_service/domain/review/dto/request/ReviewCreateRequest.java
@@ -1,0 +1,11 @@
+package com.reboot_course.review_service.domain.review.dto.request;
+
+import lombok.Builder;
+
+@Builder
+public record ReviewCreateRequest(
+        Long userId,
+        Integer score,
+        String content
+) {
+}
