+++ src/main/java/com/reboot_course/review_service/domain/review/entity/Review.java
@@ -1,0 +1,47 @@
+package com.reboot_course.review_service.domain.review.entity;
+
+
+import com.reboot_course.review_service.domain.product.entity.Product;
+import com.reboot_course.review_service.domain.user.entity.User;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
+import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@Builder
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Review {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, columnDefinition = "TINYINT")
+    @Min(0)
+    @Max(5)
+    private Integer score;
+
+    @Column(nullable = false, length = 1000)
+    private String content;
+
+    @Column
+    private String imageUrl;
+
+    @CreationTimestamp
+    @Column(nullable = false, updatable = false)
+    private LocalDateTime createdAt;
+
+    // 연관 관계
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_id", nullable = false)
+    private Product product;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
+}
