+++ src/main/java/com/reboot_course/review_service/domain/review/exception/ReviewException.java
@@ -1,0 +1,7 @@
+package com.reboot_course.review_service.domain.review.exception;
+
+public class ReviewException extends RuntimeException {
+    public ReviewException(String message) {
+        super(message);
+    }
+}
