--- src/main/java/com/reboot_course/review_service/dto/common/CommonResponse.java
@@ -1,4 +1,0 @@
-package com.reboot_course.review_service.dto.common;
-
-public record CommonResponse(String message) {
-}
