--- src/main/java/com/reboot_course/review_service/entity/Product.java
@@ -1,21 +1,0 @@
-package com.reboot_course.review_service.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Product {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Long reviewCount;
-
-    @Column(name = "score", nullable = false)
-    private Float averageScore;
-}
