+++ src/main/java/com/reboot_course/review_service/infrastructure/image/ImageManager.java
@@ -1,0 +1,57 @@
+package com.reboot_course.review_service.infrastructure.image;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.UrlResource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+import java.net.MalformedURLException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.util.Objects;
+import java.util.UUID;
+
+@Component
+public class ImageManager {
+
+    @Value("${file.upload-dir}")
+    private String uploadDir;
+
+    public String upload(MultipartFile image) {
+        if (image == null || image.isEmpty()) {
+            return null;
+        }
+
+        try {
+            Path uploadPath = Paths.get(uploadDir);
+            if (!Files.exists(uploadPath)) {
+                Files.createDirectories(uploadPath);
+            }
+
+            String filename = UUID.randomUUID() + getFileExtension(Objects.requireNonNull(image.getOriginalFilename()));
+            Path filePath = uploadPath.resolve(filename);
+            Files.copy(image.getInputStream(), filePath);
+
+            return "/images/" + filename;
+        } catch (IOException e) {
+            throw new RuntimeException("Failed to store file " + image.getOriginalFilename(), e);
+        }
+    }
+
+    public Resource findOneByName(String filename) throws MalformedURLException {
+        Path filePath = Paths.get(uploadDir).resolve(filename);
+        Resource resource = new UrlResource(filePath.toUri());
+        if (resource.exists()) {
+            return resource;
+        } else {
+            throw new RuntimeException("File not found " + filename);
+        }
+    }
+
+    private String getFileExtension(String filename) {
+        return filename.substring(filename.lastIndexOf("."));
+    }
+}
