--- src/main/java/com/reboot_course/review_service/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.reboot_course.review_service.repository;
-
-import com.reboot_course.review_service.entity.User;
-import org.springframework.data.repository.CrudRepository;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-}
